final version

+++ AAL/AAL/Heap.cpp
@@ -40,15 +40,15 @@ void Heap::push(FieldIter& begin, FieldIter& end)
 	while (i_iter > begin && j_iter->get() > val.get())
 	{
 		(*i_iter) = (*j_iter);
-		i_iter->get().setHeapIdx(std::distance(begin, i_iter));
+		i_iter->get().setHeapIdx((int)std::distance(begin, i_iter));
 		i_iter = j_iter;
-		int iter_part = (std::distance(begin, j_iter) - 1) / 2;
+		int iter_part = (int)(std::distance(begin, j_iter) - 1) / 2;
 		if (iter_part < 0) break;
 		j_iter = begin + ((std::distance(begin, j_iter) - 1) / 2);
 	}
 
 	(*i_iter) = val;
-	i_iter->get().setHeapIdx(std::distance(begin, i_iter));
+	i_iter->get().setHeapIdx((int)std::distance(begin, i_iter));
 }
 
 
@@ -60,7 +60,7 @@ void Heap::pop(FieldIter& begin, FieldIter& end)
 	std::reference_wrapper<Field> val = (*(end - 1));
 
 	(*(end - 1)) = (*begin);
-	(end - 1)->get().setHeapIdx(std::distance(begin, end - 1));
+	(end - 1)->get().setHeapIdx((int)std::distance(begin, end - 1));
 
 	FieldIter i_iter = begin;
 	FieldIter j_iter = begin + 1;
@@ -72,13 +72,13 @@ void Heap::pop(FieldIter& begin, FieldIter& end)
 
 		if (val.get() <= j_iter->get()) break;
 		(*i_iter) = (*j_iter);
-		i_iter->get().setHeapIdx(std::distance(begin, i_iter));
+		i_iter->get().setHeapIdx((int)std::distance(begin, i_iter));
 		i_iter = j_iter;
-		int iter_part = 2 * std::distance(begin, j_iter) + 1;
+		int iter_part = 2 * (int)std::distance(begin, j_iter) + 1;
 		if (iter_part >= std::distance(begin, end) - 1) break;
 		j_iter = begin + (2 * std::distance(begin, j_iter) + 1);
 	}
 
 	(*i_iter) = val;
-	i_iter->get().setHeapIdx(std::distance(begin, i_iter));
+	i_iter->get().setHeapIdx((int)std::distance(begin, i_iter));
 }
